+++ public/plugin.tsx
@@ -66,6 +66,11 @@ import {
 } from './vis_nlq/saved_object_loader';
 import { NLQVisualizationEmbeddableFactory } from './components/visualization/embeddable/nlq_vis_embeddable_factory';
 import { NLQ_VISUALIZATION_EMBEDDABLE_TYPE } from './components/visualization/embeddable/nlq_vis_embeddable';
+import {
+  ASSISTANT_INPUT_URL_SEARCH_KEY,
+  INDEX_PATTERN_URL_SEARCH_KEY,
+} from './components/visualization/text2viz';
+import { DEFAULT_DATA } from '../../../src/plugins/data/common';
 
 export const [getCoreStart, setCoreStart] = createGetterSetter<CoreStart>('CoreStart');
 
@@ -337,7 +342,7 @@ export class AssistantPlugin
         // T2Viz is only compatible with data sources that have certain agents configured
         isCompatible: async (context) => {
           // t2viz only supports selecting index pattern at the moment
-          if (context.datasetType === 'INDEX_PATTERN' && context.datasetId) {
+          if (context.datasetType === DEFAULT_DATA.SET_TYPES.INDEX_PATTERN && context.datasetId) {
             const res = await assistantServiceStart.client.agentConfigExists(
               [
                 TEXT2VEGA_RULE_BASED_AGENT_CONFIG_ID,
@@ -352,8 +357,24 @@ export class AssistantPlugin
           }
           return false;
         },
-        execute: async () => {
-          core.application.navigateToApp(TEXT2VIZ_APP_ID);
+        execute: async (context) => {
+          const url = new URL(core.application.getUrlForApp(TEXT2VIZ_APP_ID, { absolute: true }));
+          if (context.datasetId && context.datasetType === DEFAULT_DATA.SET_TYPES.INDEX_PATTERN) {
+            url.searchParams.set(INDEX_PATTERN_URL_SEARCH_KEY, context.datasetId);
+          }
+          /**
+           * TODO: the current implementation of getting query assistant input needs to be refactored
+           * once query assistant is moved to dashboard-assistant repo. Currently, there is no better
+           * way to get the input value as the query assistant is currently implemented in OSD core.
+           */
+          const queryAssistInputEle = document.getElementsByClassName('queryAssist__input')[0];
+          if (queryAssistInputEle instanceof HTMLInputElement) {
+            const input = queryAssistInputEle.value;
+            if (input) {
+              url.searchParams.set(ASSISTANT_INPUT_URL_SEARCH_KEY, input);
+            }
+          }
+          core.application.navigateToUrl(url.toString());
         },
       });
       const savedVisNLQLoader = createVisNLQSavedObjectLoader({